fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
transformers==4.44.2
torch==2.4.1
torchvision==0.19.1
pydantic==2.8.2
